requirements.txt
Flask
Flask-Cors
gunicorn
google-cloud-firestore
google-auth
requests